Listen for wallet events
