+++ go/client/cmd_chat_api_listen.go
@@ -6,7 +6,6 @@ package client
 import (
 	"encoding/json"
 	"fmt"
-	"strings"
 	"time"
 
 	"github.com/keybase/cli"
@@ -14,6 +13,7 @@ import (
 	"github.com/keybase/client/go/libkb"
 	"github.com/keybase/client/go/protocol/chat1"
 	keybase1 "github.com/keybase/client/go/protocol/keybase1"
+	"github.com/keybase/client/go/protocol/stellar1"
 	"github.com/keybase/go-framed-msgpack-rpc/rpc"
 	"golang.org/x/net/context"
 )
@@ -27,8 +27,9 @@ type CmdChatAPIListen struct {
 	// chat participants.
 	showExploding bool
 
-	showLocal    bool
-	subscribeDev bool
+	showLocal       bool
+	subscribeDev    bool
+	subscribeWallet bool
 }
 
 func newCmdChatAPIListen(cl *libcmdline.CommandLine, g *libkb.GlobalContext) cli.Command {
@@ -56,6 +57,10 @@ func newCmdChatAPIListen(cl *libcmdline.CommandLine, g *libkb.GlobalContext) cli
 				Name:  "dev",
 				Usage: "Subscribe to notifications for chat dev channels",
 			},
+			cli.BoolFlag{
+				Name:  "wallet",
+				Usage: "Subscribe to notifications for wallet events",
+			},
 		},
 	}
 }
@@ -64,6 +69,7 @@ func (c *CmdChatAPIListen) ParseArgv(ctx *cli.Context) error {
 	c.showExploding = ctx.Bool("exploding")
 	c.showLocal = ctx.Bool("local")
 	c.subscribeDev = ctx.Bool("dev")
+	c.subscribeWallet = ctx.Bool("wallet")
 	return nil
 }
 
@@ -80,17 +86,18 @@ func sendPing(cli keybase1.SessionClient) error {
 }
 
 func (c *CmdChatAPIListen) Run() error {
-	display := &chatNotificationDisplay{
-		Contextified: libkb.NewContextified(c.G()),
-		cmd:          c,
-	}
-	protocols := []rpc.Protocol{
-		chat1.NotifyChatProtocol(display),
-	}
-
 	sessionClient, err := GetSessionClient(c.G())
 	if err != nil {
 		return err
+	}
+
+	chatDisplay := newChatNotificationDisplay(c.G(), c.showLocal, c.showExploding)
+	protocols := []rpc.Protocol{
+		chat1.NotifyChatProtocol(chatDisplay),
+	}
+	if c.subscribeWallet {
+		stellarDisplay := newWalletNotificationDisplay(c.G())
+		protocols = append(protocols, stellar1.NotifyProtocol(stellarDisplay))
 	}
 
 	if err := RegisterProtocolsWithContext(protocols, c.G()); err != nil {
@@ -103,12 +110,17 @@ func (c *CmdChatAPIListen) Run() error {
 	channels := keybase1.NotificationChannels{
 		Chat:    true,
 		Chatdev: c.subscribeDev,
+		Wallet:  c.subscribeWallet,
 	}
 	if err := cli.SetNotifications(context.TODO(), channels); err != nil {
 		return err
 	}
+	ui := c.G().UI.GetTerminalUI()
+	ui.Printf("Listening for chat notifications. Config: showExploding: %v, showLocal: %v, subscribeDevChannels: %v\n", c.showExploding, c.showLocal, c.subscribeDev)
+	if c.subscribeWallet {
+		ui.Printf("Listening for wallet notifications\n")
+	}
 
-	display.errorf("Listening for chat notifications...\n")
 	for {
 		if err := sendPing(sessionClient); err != nil {
 			return fmt.Errorf("connection to service lost: error during ping: %v", err)
@@ -121,173 +133,30 @@ func (c *CmdChatAPIListen) GetUsage() libkb.Usage {
 	return libkb.Usage{}
 }
 
-type chatNotificationDisplay struct {
+type baseNotificationDisplay struct {
 	libkb.Contextified
-	cmd *CmdChatAPIListen
-	svc chatServiceHandler
 }
 
-func (d *chatNotificationDisplay) printf(fmt string, args ...interface{}) error {
+func newBaseNotificationDisplay(g *libkb.GlobalContext) *baseNotificationDisplay {
+	return &baseNotificationDisplay{
+		Contextified: libkb.NewContextified(g),
+	}
+}
+
+func (d *baseNotificationDisplay) printf(fmt string, args ...interface{}) error {
 	_, err := d.G().UI.GetTerminalUI().Printf(fmt, args...)
 	return err
 }
 
-func (d *chatNotificationDisplay) errorf(format string, args ...interface{}) error {
+func (d *baseNotificationDisplay) errorf(format string, args ...interface{}) error {
 	_, err := d.G().UI.GetTerminalUI().ErrorWriter().Write([]byte(fmt.Sprintf(format, args...)))
 	return err
 }
 
-type msgNotification struct {
-	Source     string              `json:"source,omitempty"`
-	Msg        *MsgSummary         `json:"msg,omitempty"`
-	Error      *string             `json:"error,omitempty"`
-	Pagination *chat1.UIPagination `json:"pagination,omitempty"`
-}
-
-func (d *chatNotificationDisplay) formatMessage(inMsg chat1.IncomingMessage) *Message {
-	state, err := inMsg.Message.State()
-	if err != nil {
-		errStr := err.Error()
-		return &Message{Error: &errStr}
+func (d *baseNotificationDisplay) printJSON(data interface{}) {
+	if jsonStr, err := json.Marshal(data); err != nil {
+		d.errorf("Error while marshaling JSON: %s\n", err)
+	} else {
+		d.printf("%s\n", string(jsonStr))
 	}
-
-	switch state {
-	case chat1.MessageUnboxedState_ERROR:
-		errStr := inMsg.Message.Error().ErrMsg
-		return &Message{Error: &errStr}
-	case chat1.MessageUnboxedState_VALID:
-		mv := inMsg.Message.Valid()
-		summary := &MsgSummary{
-			ID: mv.MessageID,
-			Channel: ChatChannel{
-				Name:        inMsg.Conv.Name,
-				MembersType: strings.ToLower(inMsg.Conv.MembersType.String()),
-				TopicType:   strings.ToLower(inMsg.Conv.TopicType.String()),
-				TopicName:   inMsg.Conv.Channel,
-				Public:      inMsg.Conv.Visibility == keybase1.TLFVisibility_PUBLIC,
-			},
-			Sender: MsgSender{
-				UID:        mv.SenderUID.String(),
-				DeviceID:   mv.SenderDeviceID.String(),
-				Username:   mv.SenderUsername,
-				DeviceName: mv.SenderDeviceName,
-			},
-			SentAt:              mv.Ctime.UnixSeconds(),
-			SentAtMs:            mv.Ctime.UnixMilliseconds(),
-			RevokedDevice:       mv.SenderDeviceRevokedAt != nil,
-			IsEphemeral:         mv.IsEphemeral,
-			IsEphemeralExpired:  mv.IsEphemeralExpired,
-			ETime:               mv.Etime,
-			HasPairwiseMacs:     mv.HasPairwiseMacs,
-			Content:             d.svc.convertMsgBody(mv.MessageBody),
-			AtMentionUsernames:  mv.AtMentions,
-			ChannelMention:      strings.ToLower(mv.ChannelMention.String()),
-			ChannelNameMentions: mv.ChannelNameMentions,
-		}
-		if mv.Reactions.Reactions != nil {
-			summary.Reactions = &mv.Reactions
-		}
-		return &Message{Msg: summary}
-	default:
-		return nil
-	}
-}
-
-func (d *chatNotificationDisplay) NewChatActivity(ctx context.Context, arg chat1.NewChatActivityArg) error {
-	if !d.cmd.showLocal && arg.Source == chat1.ChatActivitySource_LOCAL {
-		// Skip local message
-		return nil
-	}
-
-	activity := arg.Activity
-	typ, err := activity.ActivityType()
-	if err != nil {
-		return err
-	}
-	switch typ {
-	case chat1.ChatActivityType_INCOMING_MESSAGE:
-		inMsg := activity.IncomingMessage()
-		if !d.cmd.showExploding && inMsg.Message.IsEphemeral() {
-			// Skip exploding message
-			return nil
-		}
-		msg := d.formatMessage(inMsg)
-		if msg == nil {
-			return nil
-		}
-		notif := msgNotification{
-			Source:     strings.ToLower(arg.Source.String()),
-			Msg:        msg.Msg,
-			Error:      msg.Error,
-			Pagination: inMsg.Pagination,
-		}
-		if jsonStr, err := json.Marshal(notif); err == nil {
-			d.printf("%s\n", string(jsonStr))
-		} else {
-			d.errorf("Error while marshaling JSON: %s\n", err)
-		}
-	}
-	return nil
-}
-
-func (d *chatNotificationDisplay) ChatIdentifyUpdate(context.Context, keybase1.CanonicalTLFNameAndIDWithBreaks) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatTLFFinalize(context.Context, chat1.ChatTLFFinalizeArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatTLFResolve(context.Context, chat1.ChatTLFResolveArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatInboxStale(context.Context, keybase1.UID) error { return nil }
-func (d *chatNotificationDisplay) ChatThreadsStale(context.Context, chat1.ChatThreadsStaleArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatTypingUpdate(context.Context, []chat1.ConvTypingUpdate) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatJoinedConversation(context.Context, chat1.ChatJoinedConversationArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatLeftConversation(context.Context, chat1.ChatLeftConversationArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatResetConversation(context.Context, chat1.ChatResetConversationArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatInboxSyncStarted(context.Context, keybase1.UID) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatInboxSynced(context.Context, chat1.ChatInboxSyncedArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatSetConvRetention(context.Context, chat1.ChatSetConvRetentionArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatSetTeamRetention(context.Context, chat1.ChatSetTeamRetentionArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatSetConvSettings(context.Context, chat1.ChatSetConvSettingsArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatSubteamRename(context.Context, chat1.ChatSubteamRenameArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatKBFSToImpteamUpgrade(context.Context, chat1.ChatKBFSToImpteamUpgradeArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatAttachmentUploadStart(context.Context, chat1.ChatAttachmentUploadStartArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatAttachmentUploadProgress(context.Context, chat1.ChatAttachmentUploadProgressArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatPaymentInfo(context.Context, chat1.ChatPaymentInfoArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatRequestInfo(context.Context, chat1.ChatRequestInfoArg) error {
-	return nil
-}
-func (d *chatNotificationDisplay) ChatPromptUnfurl(context.Context, chat1.ChatPromptUnfurlArg) error {
-	return nil
 }
